requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.10.0
    # via -r requirements.in
anyio==3.6.2
    # via starlette
autopep8==2.0.1
    # via -r requirements.in
click==8.1.3
    # via uvicorn
fastapi==0.92.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
psycopg2-binary==2.9.5
    # via -r requirements.in
pycodestyle==2.10.0
    # via autopep8
pydantic==1.10.5
    # via fastapi
python-dotenv==0.21.1
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.4
    # via
    #   -r requirements.in
    #   alembic
starlette==0.25.0
    # via fastapi
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy
uvicorn==0.20.0
    # via -r requirements.in
